Add 'postgres-version' input parameter
One of the foundational principles of this action was to use
preinstalled PostgreSQL binaries. The rationale was simple:

 * Make sure the action is fast and no installation is required.
 * Make sure it's safe and doesn't reach out to external servers.
 * Make sure it's auditable and everyone can verify it contains no
   malicious code.

It's quite common, however, for applications to follow upstream
PostgreSQL releases and switch to them as soon as possible. This
applications look to install the latest stable version instead.

This patch adds a new `postgres-version` input parameter that controls
what major version of PostgreSQL to install. Please note, the parameter
receives only major part of the version, e.g. "14'. It's impossible to
request any minor release.

Co-Authored-By: Roman Podoliaka <roman.podoliaka@gmail.com>
Co-Authored-By: Ruslan Kiyanchuk <ruslan.kiyanchuk@gmail.com>

Resolves: #14